requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
pydantic==2.5.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pillow==10.1.0